Further project refactoring
