+++ machinelearning/.gitignore
@@ -5,5 +5,6 @@
 !datasets_generation.ipynb
 !datasets_analysis.ipynb
 !datasets_revision.ipynb
-!preprocessing.ipynb
+!datasets_preprocessing.ipynb
+!authors_preprocessing.ipynb
 !main.ipynb
